// src/dto/address.dto.ts
export interface IAddress {
    buildingName?: string;
    fullUnitNumber?: string;
    block?: string;
    floor?: string;
    unitNo?: string;
    streetLine1?: string;
    streetLine2?: string;
    subArea?: string;
    area?: string;
    state?: string;
    postcode?: string;
    country?: string;
    countryCode?: string;
}